+++ api/test/integration/lottery.e2e-spec.ts
@@ -1,9 +1,13 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import {
+  ApplicationStatusEnum,
+  ApplicationSubmissionTypeEnum,
+  LanguagesEnum,
   ListingsStatusEnum,
   LotteryStatusEnum,
   MultiselectQuestionsApplicationSectionEnum,
+  ReviewOrderTypeEnum,
   UnitTypeEnum,
 } from '@prisma/client';
 import request from 'supertest';
@@ -27,12 +31,31 @@ import { multiselectQuestionFactory } from '../../prisma/seed-helpers/multiselec
 import { reservedCommunityTypeFactoryAll } from '../../prisma/seed-helpers/reserved-community-type-factory';
 import { LotteryService } from '../../src/services/lottery.service';
 import { ApplicationCsvQueryParams } from 'src/dtos/applications/application-csv-query-params.dto';
+import { EmailService } from 'src/services/email.service';
 
 describe('Application Controller Tests', () => {
   let app: INestApplication;
   let prisma: PrismaService;
   let lotteryService: LotteryService;
   let cookies = '';
+  let adminAccessToken: string;
+
+  const testEmailService = {
+    /* eslint-disable @typescript-eslint/no-empty-function */
+    lotteryReleased: async () => {},
+    lotteryPublishedAdmin: async () => {},
+    lotteryPublishedApplicant: async () => {},
+  };
+
+  const mockLotteryReleased = jest.spyOn(testEmailService, 'lotteryReleased');
+  const mockLotteryPublishedAdmin = jest.spyOn(
+    testEmailService,
+    'lotteryPublishedAdmin',
+  );
+  const mockLotteryPublishedApplicant = jest.spyOn(
+    testEmailService,
+    'lotteryPublishedApplicant',
+  );
 
   const createMultiselectQuestion = async (
     jurisdictionId: string,
@@ -58,7 +81,10 @@ describe('Application Controller Tests', () => {
   beforeAll(async () => {
     const moduleFixture: TestingModule = await Test.createTestingModule({
       imports: [AppModule],
-    }).compile();
+    })
+      .overrideProvider(EmailService)
+      .useValue(testEmailService)
+      .compile();
 
     app = moduleFixture.createNestApplication();
     prisma = moduleFixture.get<PrismaService>(PrismaService);
@@ -87,6 +113,10 @@ describe('Application Controller Tests', () => {
       .expect(201);
 
     cookies = resLogIn.headers['set-cookie'];
+
+    adminAccessToken = resLogIn.header?.['set-cookie'].find((cookie) =>
+      cookie.startsWith('access-token='),
+    );
   });
 
   afterAll(async () => {
@@ -454,11 +484,11 @@ describe('Application Controller Tests', () => {
           },
         } as unknown as ExpressRequest,
         {} as Response,
-        { listingId: listing1Created.id },
+        { id: listing1Created.id },
       );
 
       const queryParams: ApplicationCsvQueryParams = {
-        listingId: listing1Created.id,
+        id: listing1Created.id,
       };
       const query = stringify(queryParams as any);
 
@@ -467,6 +497,371 @@ describe('Application Controller Tests', () => {
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .set('Cookie', cookies)
         .expect(200);
+    });
+  });
+
+  describe('expireLotteries endpoint', () => {
+    it('should only expire listing lotteries that are past due', async () => {
+      const jurisdictionA = await prisma.jurisdictions.create({
+        data: jurisdictionFactory(),
+      });
+      await reservedCommunityTypeFactoryAll(jurisdictionA.id, prisma);
+      const expiration_date = new Date();
+      expiration_date.setDate(
+        expiration_date.getDate() -
+          Number(process.env.LOTTERY_DAYS_TILL_EXPIRY || 45) -
+          1,
+      );
+      const expiredListingData = await listingFactory(
+        jurisdictionA.id,
+        prisma,
+        {
+          status: ListingsStatusEnum.closed,
+          closedAt: expiration_date,
+          reviewOrderType: ReviewOrderTypeEnum.lottery,
+        },
+      );
+      const expiredListing = await prisma.listings.create({
+        data: expiredListingData,
+      });
+
+      const recentlyClosedListingData = await listingFactory(
+        jurisdictionA.id,
+        prisma,
+        {
+          status: ListingsStatusEnum.closed,
+          closedAt: new Date(),
+          reviewOrderType: ReviewOrderTypeEnum.lottery,
+        },
+      );
+      const recentlyClosedListing = await prisma.listings.create({
+        data: recentlyClosedListingData,
+      });
+
+      const openListingData = await listingFactory(jurisdictionA.id, prisma, {
+        status: ListingsStatusEnum.active,
+        reviewOrderType: ReviewOrderTypeEnum.lottery,
+      });
+      const openListing = await prisma.listings.create({
+        data: openListingData,
+      });
+
+      const res = await request(app.getHttpServer())
+        .put(`/listings/expireLotteries`)
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .set('Cookie', adminAccessToken)
+        .expect(200);
+
+      expect(res.body.success).toEqual(true);
+
+      const postJobListing = await prisma.listings.findUnique({
+        where: {
+          id: expiredListing.id,
+        },
+      });
+
+      expect(postJobListing.lotteryStatus).toEqual(LotteryStatusEnum.expired);
+
+      const postJobListing2 = await prisma.listings.findUnique({
+        where: {
+          id: recentlyClosedListing.id,
+        },
+      });
+
+      expect(postJobListing2.lotteryStatus).toBeNull;
+
+      const postJobListing3 = await prisma.listings.findUnique({
+        where: {
+          id: openListing.id,
+        },
+      });
+
+      expect(postJobListing3.lotteryStatus).toBeNull;
+    });
+  });
+
+  describe('lottery status endpoint', () => {
+    let adminUser, adminAccessToken;
+    beforeAll(async () => {
+      adminUser = await prisma.userAccounts.create({
+        data: await userFactory({
+          roles: {
+            isAdmin: true,
+          },
+          mfaEnabled: false,
+          confirmedAt: new Date(),
+        }),
+      });
+
+      const res = await request(app.getHttpServer())
+        .post('/auth/login')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({ email: adminUser.email, password: 'Abcdef12345!' })
+        .expect(201);
+
+      adminAccessToken = res.header?.['set-cookie'].find((cookie) =>
+        cookie.startsWith('access-token='),
+      );
+    });
+
+    it("should error when trying to update listing that doesn't exist", async () => {
+      const id = randomUUID();
+      const res = await request(app.getHttpServer())
+        .put('/listings/lotteryStatus')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({
+          listingId: id,
+          lotteryStatus: LotteryStatusEnum.ran,
+        })
+        .set('Cookie', adminAccessToken)
+        .expect(404);
+      expect(res.body.message).toEqual(
+        `listingId ${id} was requested but not found`,
+      );
+    });
+
+    it('should error if user is not an admin and tries to update to ran or releasedToPartner', async () => {
+      const jurisdictionA = await prisma.jurisdictions.create({
+        data: jurisdictionFactory(),
+      });
+      await reservedCommunityTypeFactoryAll(jurisdictionA.id, prisma);
+      const listingData = await listingFactory(jurisdictionA.id, prisma, {
+        status: ListingsStatusEnum.closed,
+      });
+      const listing = await prisma.listings.create({
+        data: listingData,
+      });
+
+      await request(app.getHttpServer())
+        .put('/listings/lotteryStatus')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({
+          listingId: listing.id,
+          lotteryStatus: LotteryStatusEnum.ran,
+        })
+        .expect(403);
+
+      await request(app.getHttpServer())
+        .put('/listings/lotteryStatus')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({
+          listingId: listing.id,
+          lotteryStatus: LotteryStatusEnum.releasedToPartners,
+        })
+        .expect(403);
+    });
+
+    it('should update listing lottery status to releasedToPartners from ran', async () => {
+      const jurisdictionA = await prisma.jurisdictions.create({
+        data: jurisdictionFactory(),
+      });
+      await reservedCommunityTypeFactoryAll(jurisdictionA.id, prisma);
+      const listingData = await listingFactory(jurisdictionA.id, prisma, {
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.ran,
+      });
+      const appUpdate = new Date();
+      appUpdate.setDate(appUpdate.getDate() - 1);
+      const listing = await prisma.listings.create({
+        data: {
+          ...listingData,
+          lotteryLastRunAt: new Date(),
+          lastApplicationUpdateAt: appUpdate,
+        },
+      });
+
+      const partnerUser = await prisma.userAccounts.create({
+        data: await userFactory({
+          roles: {
+            isPartner: true,
+            isAdmin: false,
+            isJurisdictionalAdmin: false,
+          },
+          listings: [listing.id],
+          jurisdictionIds: [jurisdictionA.id],
+          confirmedAt: new Date(),
+        }),
+      });
+
+      const res = await request(app.getHttpServer())
+        .put('/listings/lotteryStatus')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({
+          listingId: listing.id,
+          lotteryStatus: LotteryStatusEnum.releasedToPartners,
+        })
+        .set('Cookie', adminAccessToken)
+        .expect(200);
+      expect(res.body.success).toEqual(true);
+
+      expect(mockLotteryReleased).toBeCalledWith(
+        expect.objectContaining({
+          email: adminUser.email,
+        }),
+        {
+          id: listing.id,
+          name: listing.name,
+          juris: expect.stringMatching(jurisdictionA.id),
+        },
+        expect.arrayContaining([partnerUser.email, adminUser.email]),
+        process.env.PARTNERS_PORTAL_URL,
+      );
+    });
+
+    it('should error trying to update listing lottery status to releasedToPartners from ran if there are new paper application updates', async () => {
+      const jurisdictionA = await prisma.jurisdictions.create({
+        data: jurisdictionFactory(),
+      });
+      await reservedCommunityTypeFactoryAll(jurisdictionA.id, prisma);
+      const listingData = await listingFactory(jurisdictionA.id, prisma, {
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.ran,
+      });
+      const lotteryLastRun = new Date();
+      lotteryLastRun.setDate(lotteryLastRun.getDate() - 1);
+      const listing = await prisma.listings.create({
+        data: {
+          ...listingData,
+          lotteryLastRunAt: lotteryLastRun,
+          lastApplicationUpdateAt: new Date(),
+        },
+      });
+
+      await request(app.getHttpServer())
+        .put('/listings/lotteryStatus')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({
+          listingId: listing.id,
+          lotteryStatus: LotteryStatusEnum.releasedToPartners,
+        })
+        .set('Cookie', adminAccessToken)
+        .expect(400);
+    });
+
+    it('should update listing lottery status to ran from releasedToPartners aka retract', async () => {
+      const jurisdictionA = await prisma.jurisdictions.create({
+        data: jurisdictionFactory(),
+      });
+      await reservedCommunityTypeFactoryAll(jurisdictionA.id, prisma);
+      const listingData = await listingFactory(jurisdictionA.id, prisma, {
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.releasedToPartners,
+      });
+      const listing = await prisma.listings.create({
+        data: listingData,
+      });
+
+      const res = await request(app.getHttpServer())
+        .put('/listings/lotteryStatus')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({
+          listingId: listing.id,
+          lotteryStatus: LotteryStatusEnum.ran,
+        })
+        .set('Cookie', adminAccessToken)
+        .expect(200);
+      expect(res.body.success).toEqual(true);
+    });
+
+    it('should update listing lottery status to publishedToPublic from releasedToPartners', async () => {
+      const jurisdictionA = await prisma.jurisdictions.create({
+        data: jurisdictionFactory(),
+      });
+      await reservedCommunityTypeFactoryAll(jurisdictionA.id, prisma);
+      const listingData = await listingFactory(jurisdictionA.id, prisma, {
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.releasedToPartners,
+        applications: [
+          {
+            preferences: [],
+            status: ApplicationStatusEnum.submitted,
+            confirmationCode: 'ABCD1234',
+            submissionType: ApplicationSubmissionTypeEnum.electronical,
+            language: LanguagesEnum.en,
+            applicant: {
+              create: {
+                emailAddress: 'applicant@email.com',
+              },
+            },
+          },
+          {
+            preferences: [],
+            status: ApplicationStatusEnum.submitted,
+            confirmationCode: 'EFGH5678',
+            submissionType: ApplicationSubmissionTypeEnum.electronical,
+            language: LanguagesEnum.es,
+            applicant: {
+              create: {
+                emailAddress: 'applicant2@email.com',
+              },
+            },
+          },
+          {
+            preferences: [],
+            status: ApplicationStatusEnum.submitted,
+            confirmationCode: 'IJKL9012',
+            submissionType: ApplicationSubmissionTypeEnum.electronical,
+            language: null,
+            applicant: {
+              create: {
+                emailAddress: 'applicant3@email.com',
+              },
+            },
+          },
+        ],
+      });
+      const listing = await prisma.listings.create({
+        data: listingData,
+      });
+
+      const partnerUser = await prisma.userAccounts.create({
+        data: await userFactory({
+          roles: {
+            isPartner: true,
+            isAdmin: false,
+            isJurisdictionalAdmin: false,
+          },
+          listings: [listing.id],
+          jurisdictionIds: [jurisdictionA.id],
+          confirmedAt: new Date(),
+        }),
+      });
+
+      const res = await request(app.getHttpServer())
+        .put('/listings/lotteryStatus')
+        .set({ passkey: process.env.API_PASS_KEY || '' })
+        .send({
+          listingId: listing.id,
+          lotteryStatus: LotteryStatusEnum.publishedToPublic,
+        })
+        .set('Cookie', adminAccessToken)
+        .expect(200);
+      expect(res.body.success).toEqual(true);
+
+      expect(mockLotteryPublishedAdmin).toBeCalledWith(
+        expect.objectContaining({
+          email: adminUser.email,
+        }),
+        {
+          id: listing.id,
+          name: listing.name,
+          juris: expect.stringMatching(jurisdictionA.id),
+        },
+        expect.arrayContaining([partnerUser.email, adminUser.email]),
+        process.env.PARTNERS_PORTAL_URL,
+      );
+
+      expect(mockLotteryPublishedApplicant).toBeCalledWith(
+        {
+          id: listing.id,
+          name: listing.name,
+          juris: expect.stringMatching(jurisdictionA.id),
+        },
+        expect.objectContaining({
+          en: ['applicant@email.com', 'applicant3@email.com'],
+          es: ['applicant2@email.com'],
+        }),
+      );
     });
   });
 });
