+++ api/test/integration/permission-tests/permission-as-admin.e2e-spec.ts
@@ -1282,7 +1282,7 @@ describe('Testing Permissioning of endpoints as Admin User', () => {
 
     it('should succeed for expireLotteries endpoint', async () => {
       await request(app.getHttpServer())
-        .put(`/listings/expireLotteries`)
+        .put(`/lottery/expireLotteries`)
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .set('Cookie', cookies)
         .expect(200);
@@ -1302,7 +1302,7 @@ describe('Testing Permissioning of endpoints as Admin User', () => {
       });
 
       await request(app.getHttpServer())
-        .put('/listings/lotteryStatus')
+        .put('/lottery/lotteryStatus')
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .send({
           listingId: listing.id,
