+++ api/test/integration/permission-tests/permission-as-juris-admin-wrong-juris.e2e-spec.ts
@@ -1105,7 +1105,7 @@ describe('Testing Permissioning of endpoints as Jurisdictional Admin in the wron
 
     it('should succeed for expireLotteries endpoint', async () => {
       await request(app.getHttpServer())
-        .put(`/listings/expireLotteries`)
+        .put(`/lottery/expireLotteries`)
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .set('Cookie', cookies)
         .expect(200);
@@ -1120,7 +1120,7 @@ describe('Testing Permissioning of endpoints as Jurisdictional Admin in the wron
       });
 
       await request(app.getHttpServer())
-        .put('/listings/lotteryStatus')
+        .put('/lottery/lotteryStatus')
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .send({
           listingId: listing.id,
