+++ api/test/integration/permission-tests/permission-as-no-user.e2e-spec.ts
@@ -1128,7 +1128,7 @@ describe('Testing Permissioning of endpoints as logged out user', () => {
 
     it('should error as forbidden for expireLotteries endpoint', async () => {
       await request(app.getHttpServer())
-        .put(`/listings/expireLotteries`)
+        .put(`/lottery/expireLotteries`)
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .set('Cookie', cookies)
         .expect(403);
@@ -1148,7 +1148,7 @@ describe('Testing Permissioning of endpoints as logged out user', () => {
       });
 
       await request(app.getHttpServer())
-        .put('/listings/lotteryStatus')
+        .put('/lottery/lotteryStatus')
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .send({
           listingId: listing.id,
