+++ api/test/integration/permission-tests/permission-as-partner-correct-listing.e2e-spec.ts
@@ -1110,7 +1110,7 @@ describe('Testing Permissioning of endpoints as partner with correct listing', (
 
     it('should error as forbidden for expireLotteries endpoint', async () => {
       await request(app.getHttpServer())
-        .put(`/listings/expireLotteries`)
+        .put(`/lottery/expireLotteries`)
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .set('Cookie', cookies)
         .expect(403);
@@ -1118,7 +1118,7 @@ describe('Testing Permissioning of endpoints as partner with correct listing', (
 
     it('should succeed for lottery status endpoint', async () => {
       await request(app.getHttpServer())
-        .put('/listings/lotteryStatus')
+        .put('/lottery/lotteryStatus')
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .send({
           listingId: closedUserListingId2,
