+++ api/test/integration/permission-tests/permission-as-partner-wrong-listing.e2e-spec.ts
@@ -1066,7 +1066,7 @@ describe('Testing Permissioning of endpoints as partner with wrong listing', () 
 
     it('should error as forbidden for expireLotteries endpoint', async () => {
       await request(app.getHttpServer())
-        .put(`/listings/expireLotteries`)
+        .put(`/lottery/expireLotteries`)
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .set('Cookie', cookies)
         .expect(403);
@@ -1074,7 +1074,7 @@ describe('Testing Permissioning of endpoints as partner with wrong listing', () 
 
     it('should error as forbidden for lottery status endpoint', async () => {
       await request(app.getHttpServer())
-        .put('/listings/lotteryStatus')
+        .put('/lottery/lotteryStatus')
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .send({
           listingId: listingId,
