+++ api/test/integration/permission-tests/permission-as-public.e2e-spec.ts
@@ -1181,7 +1181,7 @@ describe('Testing Permissioning of endpoints as public user', () => {
 
     it('should error as forbidden for expireLotteries endpoint', async () => {
       await request(app.getHttpServer())
-        .put(`/listings/expireLotteries`)
+        .put(`/lottery/expireLotteries`)
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .set('Cookie', cookies)
         .expect(403);
@@ -1201,7 +1201,7 @@ describe('Testing Permissioning of endpoints as public user', () => {
       });
 
       await request(app.getHttpServer())
-        .put('/listings/lotteryStatus')
+        .put('/lottery/lotteryStatus')
         .set({ passkey: process.env.API_PASS_KEY || '' })
         .send({
           listingId: listing.id,
